Add files via upload
Steps to running the application effectively:
1. Run the server, folder is called project-server.
2. Note that for all the other parts of the project to work, i.e. the client GUI and the
web app, the server needs to be up and running.
3. Once the server is up and running, you will notice that there are already 12 songs
that are pre-added to the server. This is to enable you to test out additional
functionalities in the client GUI and web app.
4. To run the client GUI, you need to run the project-client, folder is called projectclient.
5. Once run, a GUI will pop up with the functionalities of adding, deleting and
updating the song’s lyrics.
6. The web app can be run by clicking the index.html in the project-web-app folder.
Upon running, the html page will be opened, and you will be able to view the
lyrics that have been stored in the server.
7. Once all functionality is checked and tested, then the server can now be shut
down signifying the end of the project run.
How to run Project:
The project can be run using two ways:
1. Using an IDE, Netbeans or Intellij can be used to run the project.
2. The project can be run from the terminal as well by using the command “gradlew
bootRun” in the root of the project. Note that to do this, one must have gradle
installed in their machine.
3. The above apply to the server and client located in project-server and projectclient respectively.
4. The web app can be run by launching the index.html page on a browser located
in the project-web-app folder.